[metadata manager m8s] refactor code to utils
